# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "FeedConfig": {
        "additionalProperties": false,
        "properties": {
          "bootstrap": {
            "anyOf": [
              {
                "enum": [
                  "skip-and-mark",
                  "send-last-N",
                  "send-all"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bootstrap"
          },
          "bootstrap_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bootstrap Count"
          },
          "chat_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chat Id"
          },
          "failure": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FailureConfig"
              },
              {
                "type": "null"
              }
            ]
          },
          "llm": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LLMConfig"
              },
              {
                "type": "null"
              }
            ]
          },
          "max_llm_calls_per_run": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Hard cap on total LLM calls per run for this feed. None = unlimited (back-compat).",
            "title": "Max Llm Calls Per Run"
          },
          "name": {
            "minLength": 1,
            "pattern": "^[a-z0-9][a-z0-9._-]*$",
            "title": "Name",
            "type": "string"
          },
          "pipeline": {
            "items": {
              "$ref": "#/components/schemas/StageSpec"
            },
            "minItems": 1,
            "title": "Pipeline",
            "type": "array"
          },
          "render": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RenderConfig"
              },
              {
                "type": "null"
              }
            ]
          },
          "schedule": {
            "title": "Schedule",
            "type": "string"
          },
          "sinks": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "minItems": 1,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sinks"
          },
          "sources": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "minItems": 1,
            "title": "Sources",
            "type": "array"
          }
        },
        "required": [
          "name",
          "schedule",
          "sources",
          "pipeline"
        ],
        "title": "FeedConfig",
        "type": "object"
      },
      "FeedListResponse": {
        "additionalProperties": false,
        "description": "List view of a feed (subset of FeedConfig).",
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "pipeline_stages": {
            "items": {
              "type": "string"
            },
            "title": "Pipeline Stages",
            "type": "array"
          },
          "schedule": {
            "title": "Schedule",
            "type": "string"
          },
          "sinks_count": {
            "title": "Sinks Count",
            "type": "integer"
          },
          "sources_count": {
            "title": "Sources Count",
            "type": "integer"
          }
        },
        "required": [
          "name",
          "schedule",
          "sources_count",
          "pipeline_stages",
          "sinks_count"
        ],
        "title": "FeedListResponse",
        "type": "object"
      },
...
      "FeedStatusResponse": {
        "additionalProperties": false,
        "description": "Status of a single feed (combines config + feed_runs row).",
        "properties": {
          "alert_active": {
            "title": "Alert Active",
            "type": "boolean"
          },
          "bootstrapped": {
            "title": "Bootstrapped",
            "type": "boolean"
          },
          "consecutive_failures": {
            "title": "Consecutive Failures",
            "type": "integer"
          },
          "last_attempt_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Attempt At"
          },
          "last_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error"
          },
          "last_success_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Success At"
          },
          "llm_model": {
            "title": "Llm Model",
            "type": "string"
          },
          "llm_provider": {
            "title": "Llm Provider",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "schedule": {
            "title": "Schedule",
            "type": "string"
          }
        },
        "required": [
          "name",
          "schedule",
          "llm_provider",
          "llm_model",
          "last_success_at",
          "last_attempt_at",
          "last_error",
          "consecutive_failures",
          "alert_active",
          "bootstrapped"
        ],
        "title": "FeedStatusResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WriteResponse": {
        "additionalProperties": false,
        "description": "Generic response for write endpoints.",
        "properties": {
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ],
        "title": "WriteResponse",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Self-hosted content aggregation daemon \u2014 REST API",
    "title": "glean",
    "version": "1.4.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/config/feeds": {
      "get": {
        "operationId": "list_feeds_api_v1_config_feeds_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-Glean-Api-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Glean-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FeedListResponse"
                  },
                  "title": "Response List Feeds Api V1 Config Feeds Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Feeds",
        "tags": [
          "config"
        ]
      },
      "post": {
        "operationId": "create_feed_api_v1_config_feeds_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Glean-Api-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Glean-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedConfig"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WriteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Feed",
        "tags": [
          "config"
        ]
      }
    },
    "/api/v1/config/feeds/{name}": {
      "delete": {
        "operationId": "delete_feed_api_v1_config_feeds__name__delete",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Glean-Api-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Glean-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WriteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Feed",
        "tags": [
          "config"
        ]
      },
      "get": {
        "operationId": "get_feed_api_v1_config_feeds__name__get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Glean-Api-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Glean-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedConfig"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Feed",
        "tags": [
          "config"
        ]
      },
      "put": {
        "operationId": "update_feed_api_v1_config_feeds__name__put",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Glean-Api-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Glean-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedConfig"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WriteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Feed",
        "tags": [
          "config"
        ]
      }
    },
...
    "/api/v1/feeds": {
      "get": {
        "operationId": "list_feeds_api_v1_feeds_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-Glean-Api-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Glean-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FeedStatusResponse"
                  },
                  "title": "Response List Feeds Api V1 Feeds Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Feeds",
        "tags": [
          "feeds"
        ]
      }
    },
...
    "/api/v1/feeds/{name}/run": {
      "post": {
        "description": "Run a feed off-schedule with real sends.",
        "operationId": "run_feed_now_api_v1_feeds__name__run_post",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Glean-Api-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Glean-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResultResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Feed Now",
        "tags": [
          "feeds"
        ]
      }
    },
    "/api/v1/feeds/{name}/runs": {
      "get": {
        "operationId": "list_feed_runs_api_v1_feeds__name__runs_get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "before",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Before"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "success",
                    "failure",
                    "skip"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "header",
            "name": "X-Glean-Api-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Glean-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FeedRunResponse"
                  },
                  "title": "Response List Feed Runs Api V1 Feeds  Name  Runs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Feed Runs",
        "tags": [
          "feeds"
        ]
      }
    },
    "/api/v1/feeds/{name}/status": {
      "get": {
        "operationId": "feed_status_api_v1_feeds__name__status_get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Glean-Api-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Glean-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Feed Status",
        "tags": [
          "feeds"
        ]
      }
    },
    "/api/v1/feeds/{name}/suppressed": {
      "get": {
        "operationId": "list_feed_suppressed_api_v1_feeds__name__suppressed_get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "before",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Before"
            }
          },
          {
            "in": "header",
            "name": "X-Glean-Api-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Glean-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FeedSuppressedResponse"
                  },
                  "title": "Response List Feed Suppressed Api V1 Feeds  Name  Suppressed Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Feed Suppressed",
        "tags": [
          "feeds"
        ]
      }
    },
    "/api/v1/feeds/{name}/test": {
      "post": {
        "description": "Dry-run a feed (no Telegram, no state writes for sending).",
        "operationId": "test_feed_api_v1_feeds__name__test_post",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Glean-Api-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Glean-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResultResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Test Feed",
        "tags": [
          "feeds"
        ]
      }
    },
...
  }
}